Demonstrate how to remove static data for gpio
When only one instance of a device exists, we can completely remove the
static data structures and replace them with constant macros. This
demonstrates how to do this for the sifive,gpio0 device.

Additional optimizations can take place when a given value is constant
across all instances of a device.

Signed-off-by: Nathaniel Graff <nathaniel.graff@sifive.com>